重构：聊天视图-拆分聊天视图状态与交互逻辑
  - 抽离 AI 润色、表情面板及 @/命令建议逻辑
  - 独立消息多选、转发、复制、归档与重排流程
  - 封装输入/编辑实时预览、身份差分和拖拽排序
  - 降低 chat.vue 单文件复杂度，保持现有交互接线
